Start cleaning up some of the ingestor tests
