+++ catalogue_pipeline/ingestor/src/test/scala/uk/ac/wellcome/platform/ingestor/services/IngestorWorkerServiceTest.scala
@@ -8,17 +8,14 @@ import org.scalatest.{Assertion, FunSpec, Matchers}
 import uk.ac.wellcome.elasticsearch.ElasticCredentials
 import uk.ac.wellcome.elasticsearch.test.fixtures.ElasticsearchFixtures
 import uk.ac.wellcome.json.JsonUtil._
-import uk.ac.wellcome.messaging.message.MessageStream
 import uk.ac.wellcome.messaging.test.fixtures.SQS.QueuePair
 import uk.ac.wellcome.messaging.test.fixtures.{Messaging, SQS}
 import uk.ac.wellcome.models.work.generators.WorksGenerators
 import uk.ac.wellcome.models.work.internal.{IdentifiedBaseWork, IdentifierType, Subject}
-import uk.ac.wellcome.platform.ingestor.config.models.{IngestElasticConfig, IngestorConfig}
-import uk.ac.wellcome.platform.ingestor.fixtures.WorkIndexerFixtures
+import uk.ac.wellcome.platform.ingestor.fixtures.{WorkIndexerFixtures, WorkerServiceFixture}
 import uk.ac.wellcome.test.fixtures.TestWith
 
 import scala.concurrent.ExecutionContext.Implicits.global
-import scala.concurrent.duration._
 
 class IngestorWorkerServiceTest
     extends FunSpec
@@ -27,6 +24,7 @@ class IngestorWorkerServiceTest
     with Messaging
     with ElasticsearchFixtures
     with SQS
+    with WorkerServiceFixture
     with WorkIndexerFixtures
     with WorksGenerators
     with CustomElasticsearchMapping {
@@ -270,91 +268,36 @@ class IngestorWorkerServiceTest
   }
 
   it("returns a failed Future if indexing into Elasticsearch fails") {
-    withActorSystem { actorSystem =>
-      withMetricsSender(actorSystem) { metricsSender =>
-        withLocalSqsQueueAndDlq {
-          case QueuePair(queue, dlq) =>
-            withMessageStream[IdentifiedBaseWork, Assertion](
-              actorSystem,
-              queue,
-              metricsSender) { messageStream =>
-              val brokenRestClient: RestClient = RestClient
-                .builder(new HttpHost("localhost", 9800, "http"))
-                .setHttpClientConfigCallback(
-                  new ElasticCredentials("elastic", "changeme"))
-                .build()
+    withLocalSqsQueueAndDlq { case QueuePair(queue, dlq) =>
+        val brokenRestClient: RestClient = RestClient
+          .builder(new HttpHost("localhost", 9800, "http"))
+          .setHttpClientConfigCallback(
+            new ElasticCredentials("elastic", "changeme"))
+          .build()
 
-              val brokenElasticClient: HttpClient =
-                HttpClient.fromRestClient(brokenRestClient)
+        val brokenElasticClient: HttpClient =
+          HttpClient.fromRestClient(brokenRestClient)
 
-              val brokenWorkIndexer = new WorkIndexer(
-                elasticClient = brokenElasticClient
-              )
+        withWorkerService(queue, indexName = "works-v1", elasticClient = brokenElasticClient) { _ =>
+          val work = createIdentifiedWork
 
-              withIngestorWorkerService[Assertion](
-                indexName = "works-v1",
-                brokenWorkIndexer,
-                messageStream) { _ =>
-                val work = createIdentifiedWork
+          sendMessage[IdentifiedBaseWork](queue = queue, obj = work)
 
-                sendMessage[IdentifiedBaseWork](queue = queue, obj = work)
-
-                eventually {
-                  assertQueueEmpty(queue)
-                  assertQueueHasSize(dlq, 1)
-                }
-              }
-            }
+          eventually {
+            assertQueueEmpty(queue)
+            assertQueueHasSize(dlq, 1)
+          }
         }
-      }
     }
   }
 
   private def withIngestorWorkerService[R](indexName: String)(
-    testWith: TestWith[QueuePair, R]): R = {
-    withActorSystem { actorSystem =>
-      withMetricsSender(actorSystem) { metricsSender =>
-        withLocalSqsQueueAndDlqAndTimeout(10) {
-          case queuePair @ QueuePair(queue, dlq) =>
-            withWorkIndexer { workIndexer =>
-              withMessageStream[IdentifiedBaseWork, R](
-                actorSystem,
-                queue,
-                metricsSender) { messageStream =>
-                withIngestorWorkerService[R](
-                  indexName,
-                  workIndexer,
-                  messageStream) { _ =>
-                  testWith(queuePair)
-                }
-              }
-            }
+    testWith: TestWith[QueuePair, R]): R =
+    withLocalSqsQueueAndDlqAndTimeout(10) {
+      case queuePair@QueuePair(queue, dlq) =>
+        withWorkerService(queue, indexName = indexName) { _ =>
+          testWith(queuePair)
         }
-      }
     }
-  }
-
-  private def withIngestorWorkerService[R](
-    indexName: String,
-    workIndexer: WorkIndexer,
-    messageStream: MessageStream[IdentifiedBaseWork])(
-    testWith: TestWith[IngestorWorkerService, R]): R = {
-
-    val ingestorConfig = IngestorConfig(
-      batchSize = 100,
-      flushInterval = 5 seconds,
-      elasticConfig = IngestElasticConfig(
-        documentType = documentType,
-        indexName = indexName
-      )
-    )
-
-    val ingestorWorkerService = new IngestorWorkerService(
-      ingestorConfig = ingestorConfig,
-      elasticClient = elasticClient,
-      messageStream = messageStream
-    )
-
-    testWith(ingestorWorkerService)
   }
 }
